Make bevy_mesh optional for bevy_animate
