+++ crates/bevy_animation/src/lib.rs
@@ -13,6 +13,8 @@ pub mod animatable;
 pub mod animation_curves;
 pub mod gltf_curves;
 pub mod graph;
+#[cfg(feature = "bevy_mesh")]
+mod morph;
 pub mod transition;
 
 mod animation_event;
@@ -1249,9 +1251,12 @@ impl Plugin for AnimationPlugin {
                     // it to its own system set after `Update` but before
                     // `PostUpdate`. For now, we just disable ambiguity testing
                     // for this system.
+                    #[cfg(feature = "bevy_mesh")]
                     animate_targets
                         .before(bevy_mesh::InheritWeightSystems)
                         .ambiguous_with_all(),
+                    #[cfg(not(feature = "bevy_mesh"))]
+                    animate_targets.ambiguous_with_all(),
                     trigger_untargeted_animation_events,
                     expire_completed_transitions,
                 )
