+++ crates/bevy_internal/Cargo.toml
@@ -212,6 +212,9 @@ animation = [
 # Enables support for morph target weights in bevy_mesh
 morph = ["bevy_mesh?/morph", "bevy_render?/morph"]
 
+# Enables bevy_mesh and bevy_animation morph weight support
+morph_animation = ["morph", "bevy_animation?/bevy_mesh"]
+
 bevy_shader = ["dep:bevy_shader"]
 bevy_image = ["dep:bevy_image", "bevy_color", "bevy_asset"]
 bevy_sprite = ["dep:bevy_sprite", "bevy_camera"]
