+++ docs/cargo_features.md
@@ -53,6 +53,7 @@ The default feature set enables most of the expected features of a game engine, 
 |hdr|HDR image format support|
 |ktx2|KTX2 compressed texture support|
 |morph|Enables support for morph target weights in bevy_mesh|
+|morph_animation|Enables bevy_mesh and bevy_animation morph weight support|
 |multi_threaded|Enables multithreaded parallelism in the engine. Disabling it forces all engine tasks to run on a single thread.|
 |png|PNG image format support|
 |reflect_auto_register|Enable automatic reflect registration|
